List of changes: 1)keep the account name unique under user_id and financial combination. 2)bank statement upload controller addition to support bulk upload. 3)making the decimal to 2 places for all amounts only for quantity to 4 places. 4)Added the globalBatchOperations table to check any batch job is running or not for user_id and financial_year combination. 5)change the routes to accept more memory only for bank upload route,other routes will not allow more memory. 6)trigger of update balance is checking batch job is running or not,if yes skip the balance calculation. 7)changed the trail balance logic to include cash entries and opening balances of all accounts while calculation of trail balance. 8)cash entry to support group_id from now on. 9)entries to support narration logic creation for party,taxes and type account like purchase,sale etc.. 10)amount is converted parseFloat before summation in entries. 11)ledger to support previous balance as pagination is causing the issue,so for every page sending the total balance for that page. 12)cash controller to support view instead of reading from cash_entries as now new table is created for cashEntriesBatch,so getting consolidate data from view. 13)create a new combined_cash_entries view for cash_entries and cashEntriesBatch. 14)create-insert-initial-data-function need to balance default entry after selecting the new financial_year. 15)Add transaction id to cash_entries. 16)Added the Suspense Account to default account_group with Suspense group and added to default_account_list and default_group_list. 17)removed the description column from default account list as it renamed to gstno in account_list. 18)Added 3 dependencies multer to support the buffer reader from file,moment-timezone to support timezone in bank upload and pdfreader to support pdf page to parse the data.

+++ src/models/account.model.js
@@ -17,12 +17,12 @@ module.exports = (sequelize, Sequelize) => {
             type: Sequelize.INTEGER,
         },
         credit_balance: {
-            type: Sequelize.FLOAT,
+            type: Sequelize.DECIMAL(15, 2),
             defaultValue: 0,
             allowNull: false,
         },
         debit_balance: {
-            type: Sequelize.FLOAT,
+            type: Sequelize.DECIMAL(15, 2),
             defaultValue: 0,
             allowNull: false,
         },
